+++ packages/plugin-react-query/src/generators/infiniteQueryGenerator.tsx
@@ -175,6 +175,8 @@ export const infiniteQueryGenerator = createReactGenerator<PluginReactQuery>({
               pathParamsType={options.pathParamsType}
               dataReturnType={options.client.dataReturnType || 'data'}
               cursorParam={infiniteOptions.cursorParam}
+              nextParam={infiniteOptions.nextParam}
+              previousParam={infiniteOptions.previousParam}
               initialPageParam={infiniteOptions.initialPageParam}
               queryParam={infiniteOptions.queryParam}
             />